[FIX] Teams Adaptive Cards: Move the "actions" outside the body in the payload. Also allow refactor to allow param for the openUrl Action to work.
The request was returning 400 Bad Request because the "actions" key was being sent
inside the body. It must be sent a the same level as body instead.

See the examples on https://adaptivecards.io/explorer/Action.OpenUrl.html

or

See the examples on https://adaptivecards.io/explorer/Action.Submit.html